Refactor useItemFilter and matchItem
