+++ client/packages/common/src/utils/index.ts
@@ -12,6 +12,7 @@ export * from './object';
 export * from './types';
 export * from './files';
 export * from './BarcodeScannerContext';
+export * from './item';
 
 // having issues with tree shaking lodash
 // so we're just importing the functions we need
