+++ client/packages/requisitions/src/RequestRequisition/api/hooks/line/useRequestLines.ts
@@ -1,34 +1,12 @@
 import { useMemo } from 'react';
-import {
-  SortUtils,
-  RegexUtils,
-  useUrlQuery,
-  ItemNode,
-} from '@openmsupply-client/common';
+import { SortUtils, ItemUtils } from '@openmsupply-client/common';
 import { useRequestColumns } from '../../../DetailView/columns';
 import { useHideOverStocked } from '../index';
 import { useRequestFields } from '../document/useRequestFields';
 
-const useItemFilter = () => {
-  const { urlQuery, updateQuery } = useUrlQuery({ skipParse: ['codeOrName'] });
-  return {
-    itemFilter: urlQuery.codeOrName ?? '',
-    setItemFilter: (itemFilter: string) =>
-      updateQuery({ codeOrName: itemFilter }),
-  };
-};
-
-const matchItem = (itemFilter: string, { name, code }: Partial<ItemNode>) => {
-  const filter = RegexUtils.escapeChars(itemFilter);
-  return (
-    RegexUtils.includes(filter, name ?? '') ||
-    RegexUtils.includes(filter, code ?? '')
-  );
-};
-
 export const useRequestLines = () => {
   const { on } = useHideOverStocked();
-  const { itemFilter, setItemFilter } = useItemFilter();
+  const { itemFilter, setItemFilter } = ItemUtils.itemFilter();
   const { columns, onChangeSortBy, sortBy } = useRequestColumns();
   const { lines, minMonthsOfStock } = useRequestFields([
     'lines',
@@ -49,10 +27,10 @@ export const useRequestLines = () => {
         item =>
           item.itemStats.availableStockOnHand <
             item.itemStats.averageMonthlyConsumption * minMonthsOfStock &&
-          matchItem(itemFilter, item.item)
+          ItemUtils.matchItem(itemFilter, item.item)
       );
     } else {
-      return sorted.filter(item => matchItem(itemFilter, item.item));
+      return sorted.filter(item => ItemUtils.matchItem(itemFilter, item.item));
     }
   }, [sortBy.key, sortBy.isDesc, lines, on, minMonthsOfStock, itemFilter]);
